backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
cassandra-driver>=3.29.1
websockets==12.0
pyasyncore>=1.0.4